Use maximum concurrency for interactive runs
Similarly that the user can start multiple interactive sessions and run
code in as many kernels as the system allows, their should not be an
artificial cap on the number of concurrent interactive runs.

You can still only have one interactive run per pipeline. This is a
constraint through the UI, calling the orchest-api would actually allow
you to run multiple for the same pipeline.