Use '__name__' as 'import_name' for Flask
By using `__name__` we have higher confidence that the Flask application
(and extensions) work correctly. The only downside is the name of the
package in the logging (which will not be `app`), however, this should
not be a problem at this point since logs of different services are not
bundled.

The decision was led by the information on:
https://blog.miguelgrinberg.com/post/why-do-we-pass-name-to-the-flask-class