Minimally change pipeline definition for diffs
Since the pipeline definition is stored in JSON and JSON is unordered,
every change to the pipeline definition could end up in large diffs.
Instead of relying on the implicit order of keys in dictionaries,
especially since the pipeline definition is also interacted with in
Javascript, explicitly sort the keys.

Caveat: The order given by the user of the defined parameters is now
also automatically sorted.