Make update pull minimal set of required images
This is a change in design as it used to be the sole task of `orchest
install` to pull new images and for `orchest update` to only pull newer
versions of existing images. However, this does not work in the case
that a new required image is added to Orchest. Because one would expect
to be able to call `orchest update` but then misses a required image.