Add preserve_base_path service field
This is necessary because some services (applications) use relative
paths while others do not when it comes to handling, for example, web
client web requets. This field allows to specify if, when proxying the
service, the base path should be kept or not. For applications using
relative paths it should not be preserved.