Workaround apsscheduler and eventlet issue
This is needed because it looks like apsscheduler is not getting control
until the eventlet webserver is actually idle. In this particular case,
time.sleep(1) yields control to other calls made to the webserver, but
not to the scheduler. The scheduler  will be able to schedule the
StopInteractiveSession background task only after docker_images_rm_safe
is done trying and failed, since any kernel container from the session
would still be running.