Don't touch pipeline definition from environments view directly
This partly reverts changes from: https://github.com/orchest/orchest/pull/449

The only view to directly change the pipeline definition is the pipeline
edit view. Note that we should not rely on deletion of an environment
when the environment is deleted through the UI, because when copying a
pipeline definition it may very well be that the defined environment
does not actually exist.