Display changed config values that require an Orchest restart
This handles, from a user POV, the edge case where in one change both a
static and dynamic config var is changed. Previously the message would
have stated a restart is required, but a user might only care about the
config var that is dynamic and would end up restarting Orchest without
it being required.