Remove ability to update params through SDK
It does not make sense to expose functionality to update params through
code because every pipeline run in a job get's its own project
directory copy and so changing parameters inside jobs doesn't have any
effect to begin with. The only use case that is left is using code to
define and update parameters and deleting that code again, instead of
using the UI to define the parameter.

Due to how jobs are set up, there is no use case for dynamically
updating parameters.