Only attempt once to delete dangling images on app initizaliation
In the edge case where a dangling image could not be deleted (e.g. the
parent image failed to start but is never removed using
`docker rm $(docker ps -a -q)`) we would sleep for 1 second and retry
this 10 times. This would case the orchest-api to be unresponsive and
thereby making the FE feel unresponsive for views (e.g. `/projects`)
that query the orchest-api.

Ideally the host machine has a clean docker environment, i.e. no
dangling images, but we can't guarantee this. Luckily, this edge case
should be rare for users and only happen for developers of Orchest.