Prepare Flask services to support multiple gunicorn workers
In order to support multiple gunicorn workers the `create_app` function
can no longer handle DB migrations/upgrades and it is factored out
before the application is started.

In addition, because every gunicorn worker needs its own background
scheduler this should not lead to multiple executions of the same job
(since it is duplicated for every worker). To prevent this the jobs read
from the DB to make sure only one will end up running.

In this process the `is_werkzeug_parent` function was removed since it
was very confusing to work with. Its substitute is
`is_running_from_reloader`.