Isolate ENV vars to step level docs section
If you set ENV vars for the project or pipeline then all steps will
search in the new paths for the python environment. This could result in
failure because different steps could use different python environments
yet now they would all point to the same environment which might not
exist.